backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
supabase==2.18.1
langchain==0.3.27
langchain-openai==0.3.32
python-jose==3.4.0
